Abstract notification logic outside of the VideoService
